requirements.txt
Flask
cs50
Flask-Session
country_list
RPi.GPIO
Adafruit-DHT

